Add files via upload
Introduction:
This dataset 'Mushroom-Classification' includes descriptions of hypothetical samples corresponding to 23 species of gilled mushrooms in the Agaricus and Lepiota Family Mushroom drawn from The Audubon Society Field Guide to North American Mushrooms (1981). Each species is identified as definitely edible, definitely poisonous, or of unknown edibility and not recommended. This latter class was combined with the poisonous one. The Guide clearly states that there is no simple rule for determining the edibility of a mushroom; no rule like "leaflets three, let it be'' for Poisonous Oak and Ivy. We separated the dataset into two divisions (X,Y) for the purpose of creating these models: training and testing. When we understood about the dataset, we were required to do some pre-processing. In order to implement the algorithm and produce the best result, the dataset has been pre-processed according to particular criteria. We looked at the data's correlation and directionality before estimating the model. Then, by comparing training and testing data, assess the model's validity and utility. We used some major regression analysis to evaluate the more accurate rate. We used K-Nearest-Neighbour and Decision Forest Tree algorithms for classifying and training the model and other methods have been used to plot the correlation between variables.